refactor(rest): separate protobuf utils from grpc utils
