+++ google/cloud/google_cloud_cpp_protobuf_utils.bzl
@@ -1,0 +1,25 @@
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
+
+"""Automatically generated source lists for google_cloud_cpp_protobuf_utils - DO NOT EDIT."""
+
+google_cloud_cpp_protobuf_utils_hdrs = [
+    "internal/log_wrapper_helpers.h",
+]
+
+google_cloud_cpp_protobuf_utils_srcs = [
+    "internal/log_wrapper_helpers.cc",
+]
