+++ google/cloud/internal/log_wrapper_test.cc
@@ -61,143 +61,6 @@ google::spanner::v1::Mutation MakeMutation() {
   return mutation;
 }
 
-TEST(LogWrapper, DefaultOptions) {
-  TracingOptions tracing_options;
-  // clang-format off
-  std::string const text =
-      R"pb(google.spanner.v1.Mutation { )pb"
-      R"pb(insert { )pb"
-      R"pb(table: "Singers" )pb"
-      R"pb(columns: "SingerId" )pb"
-      R"pb(columns: "FirstName" )pb"
-      R"pb(columns: "LastName" )pb"
-      R"pb(values { values { string_value: "1" } )pb"
-      R"pb(values { string_value: "test-fname-1" } )pb"
-      R"pb(values { string_value: "test-lname-1" } } )pb"
-      R"pb(values { values { string_value: "2" } )pb"
-      R"pb(values { string_value: "test-fname-2" } )pb"
-      R"pb(values { string_value: "test-lname-2" } )pb"
-      R"pb(} )pb"
-      R"pb(} )pb"
-      R"pb(})pb";
-  // clang-format on
-  EXPECT_EQ(text, DebugString(MakeMutation(), tracing_options));
-}
-
-TEST(LogWrapper, MultiLine) {
-  TracingOptions tracing_options;
-  tracing_options.SetOptions("single_line_mode=off");
-  // clang-format off
-  std::string const text = R"pb(google.spanner.v1.Mutation {
-  insert {
-    table: "Singers"
-    columns: "SingerId"
-    columns: "FirstName"
-    columns: "LastName"
-    values {
-      values {
-        string_value: "1"
-      }
-      values {
-        string_value: "test-fname-1"
-      }
-      values {
-        string_value: "test-lname-1"
-      }
-    }
-    values {
-      values {
-        string_value: "2"
-      }
-      values {
-        string_value: "test-fname-2"
-      }
-      values {
-        string_value: "test-lname-2"
-      }
-    }
-  }
-})pb";
-  // clang-format on
-  EXPECT_EQ(text, DebugString(MakeMutation(), tracing_options));
-}
-
-TEST(LogWrapper, Truncate) {
-  TracingOptions tracing_options;
-  tracing_options.SetOptions("truncate_string_field_longer_than=8");
-  // clang-format off
-  std::string const text =
-            R"pb(google.spanner.v1.Mutation { )pb"
-            R"pb(insert { )pb"
-            R"pb(table: "Singers" )pb"
-            R"pb(columns: "SingerId" )pb"
-            R"pb(columns: "FirstNam...<truncated>..." )pb"
-            R"pb(columns: "LastName" )pb"
-            R"pb(values { values { string_value: "1" } )pb"
-            R"pb(values { string_value: "test-fna...<truncated>..." } )pb"
-            R"pb(values { string_value: "test-lna...<truncated>..." } } )pb"
-            R"pb(values { values { string_value: "2" } )pb"
-            R"pb(values { string_value: "test-fna...<truncated>..." } )pb"
-            R"pb(values { string_value: "test-lna...<truncated>..." } )pb"
-            R"pb(} )pb"
-            R"pb(} )pb"
-            R"pb(})pb";
-  // clang-format on
-  EXPECT_EQ(text, DebugString(MakeMutation(), tracing_options));
-}
-
-TEST(LogWrapper, TruncateString) {
-  TracingOptions tracing_options;
-  tracing_options.SetOptions("truncate_string_field_longer_than=8");
-  struct Case {
-    std::string actual;
-    std::string expected;
-  } cases[] = {
-      {"1234567", "1234567"},
-      {"12345678", "12345678"},
-      {"123456789", "12345678...<truncated>..."},
-      {"1234567890", "12345678...<truncated>..."},
-  };
-  for (auto const& c : cases) {
-    EXPECT_EQ(c.expected, DebugString(c.actual, tracing_options));
-  }
-}
-
-TEST(LogWrapper, Duration) {
-  google::protobuf::Duration duration;
-  duration.set_seconds((11 * 60 + 22) * 60 + 33);
-  duration.set_nanos(123456789);
-  std::string const expected =
-      R"(google.protobuf.Duration { "11h22m33.123456789s" })";
-  EXPECT_EQ(expected, DebugString(duration, TracingOptions{}.SetOptions(
-                                                "single_line_mode=on")));
-}
-
-TEST(LogWrapper, Timestamp) {
-  google::protobuf::Timestamp timestamp;
-  timestamp.set_seconds(1658470436);
-  timestamp.set_nanos(123456789);
-  std::string const expected = R"(google.protobuf.Timestamp {
-  "2022-07-22T06:13:56.123456789Z"
-})";
-  EXPECT_EQ(expected, DebugString(timestamp, TracingOptions{}.SetOptions(
-                                                 "single_line_mode=off")));
-}
-
-TEST(LogWrapper, FutureStatus) {
-  struct Case {
-    std::future_status actual;
-    std::string expected;
-  } cases[] = {
-      {std::future_status::deferred, "deferred"},
-      {std::future_status::timeout, "timeout"},
-      {std::future_status::ready, "ready"},
-  };
-  for (auto const& c : cases) {
-    EXPECT_EQ(c.expected, DebugFutureStatus(c.actual));
-  }
-}
-
 TEST(LogWrapper, StatusDetails) {
   google::rpc::ResourceInfo resource_info;
   resource_info.set_resource_type(
